[FIX] naming fix
